@@ -1,12 +1,12 @@
 name: giphy_selector
 description: A Flutter plugin enabling the selection of a GIPHY GIF, Sticker or Emoji
-version: 1.2.2
+version: 1.3.0
 homepage: https://github.com/amantoux/giphy_selector
 
 
 environment:
-  sdk: ">=2.18.0 <3.0.0"
-  flutter: ">=1.17.0"
+  sdk: ">=2.18.0 <4.0.0"
+  flutter: ">=3.0.0"
 
 dependencies:
   flutter:
